fix_Makefile

+++ ft_memcpy.c
@@ -1,10 +1,10 @@
 #include "libft.h"
 
-void    *ft_memcpy(void *restrict dest, const void *restrict src, int n)
+void    *ft_memcpy(void *restrict dest, const void *restrict src, size_t n)
 {
     if(!src)
         return(0);
-    int i;
+    size_t i;
 
     i = 0;
     while (i < n)
