+++ ft_memmove.c
@@ -1,10 +1,10 @@
 #include "libft.h"
 
-void *ft_memmove(void *dest, const void *src, int n)
+void *ft_memmove(void *dest, const void *src, size_t n)
 {
     if(!src)
         return(0);
-    int i;
+    size_t i;
 
     i = 0;
     while (i < n)
