+++ ft_strnstr.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 #include "libft.h"
-char    *ft_strnstr(const char	*big, const char *little, int len)
+char    *ft_strnstr(const char	*big, const char *little, size_t len)
 {
-    int i;
-	int j;
+    size_t i;
+	size_t j;
 	
 	i = 0;
 	j = 0;
